Bug Fixes
Bug Fixes

+++ src/chainrepublik/network/packets/misc/CDelVotePayload.java
@@ -3,6 +3,7 @@ package chainrepublik.network.packets.misc;
 import chainrepublik.kernel.UTILS;
 import chainrepublik.network.packets.CPayload;
 import chainrepublik.network.packets.blocks.CBlockPayload;
+import java.sql.ResultSet;
 
 public class CDelVotePayload extends CPayload
 {
@@ -53,6 +54,30 @@ public class CDelVotePayload extends CPayload
             !this.type.equals("ID_DOWN"))
         throw new Exception("Invalid type");
         
+        // Already voted ?
+        ResultSet rs=UTILS.DB.executeQuery("SELECT * "
+                                           + "FROM del_votes "
+                                          + "WHERE adr='"+this.target_adr+"' "
+                                            + "AND delegate='"+this.delegate+"' "
+                                            + "AND type='"+this.type+"'");
+        
+        // Has data ?
+        if (UTILS.DB.hasData(rs))
+            throw new Exception("You aready voted this delegate"); 
+        
+        // Max 10 votes
+        rs=UTILS.DB.executeQuery("SELECT COUNT(*) AS total "
+                                 + "FROM del_votes "
+                                + "WHERE adr='"+this.target_adr+"'");
+        
+        // Next
+        rs.next();
+        
+        // Max 10 votes ?
+        if (rs.getLong("total")>=10)
+             throw new Exception("Already 10 votes");
+        
+        // Hash
         String h=UTILS.BASIC.hash(this.getHash()+
 			          this.delegate+
 			          this.type);
@@ -67,11 +92,17 @@ public class CDelVotePayload extends CPayload
         // Superclass
 	super.commit(block);
         
-        // Remove old vote
+        // Already voted ?
+        ResultSet rs=UTILS.DB.executeQuery("SELECT * "
+                                           + "FROM del_votes "
+                                          + "WHERE delegate='"+this.delegate+"' "
+                                            + "AND adr='"+this.target_adr+"'");
+        
+        // Has data
+        if (UTILS.DB.hasData(rs))
         UTILS.DB.executeUpdate("DELETE FROM del_votes "
                                    + "WHERE adr='"+this.target_adr+"'");
-        
-        // Insert new vote
+        else
         UTILS.DB.executeUpdate("INSERT INTO del_votes "
                                      + "SET delegate='"+this.delegate+"', "
                                          + "adr='"+this.target_adr+"', "
